Add growth policies

+++ include/foonathan/array/block_storage_new.hpp
@@ -8,6 +8,7 @@
 #include <new>
 #include <type_traits>
 
+#include <foonathan/array/growth_policy.hpp>
 #include <foonathan/array/memory_block.hpp>
 #include <foonathan/array/raw_storage.hpp>
 
@@ -26,11 +27,17 @@ namespace foonathan
             ::operator delete[](block.memory);
         }
 
+        /// A `BlockStorage` that uses `operator new` for memory allocations.
+        ///
+        /// It does not have a small buffer optimization and uses the specified growth policy.
+        template <class GrowthPolicy>
         class block_storage_new
         {
         public:
             static constexpr auto max_capacity = memory_block::max_size();
             using exact_capacity               = std::true_type;
+
+            block_storage_new() noexcept = default;
 
             explicit block_storage_new(size_type min_initial_capacity)
             : block_(new_block(min_initial_capacity))
@@ -39,7 +46,8 @@ namespace foonathan
 
             ~block_storage_new() noexcept
             {
-                delete_block(std::move(block_));
+                if (block_)
+                    delete_block(std::move(block_));
             }
 
             block_storage_new(const block_storage_new&) = delete;
@@ -48,14 +56,16 @@ namespace foonathan
             template <typename T>
             T* reserve(size_type min_additional, T* begin_constructed, T* end_constructed)
             {
-                auto bigger = new_block(block_.size + min_additional);
+                auto bigger = new_block(GrowthPolicy::growth_size(block_.size, min_additional));
                 return move_elements(begin_constructed, end_constructed, bigger);
             }
 
             template <typename T>
             T* shrink_to_fit(T* begin_constructed, T* end_constructed)
             {
-                auto smaller = new_block(size_type(end_constructed - begin_constructed));
+                auto smaller = new_block(
+                    GrowthPolicy::shrink_size(block_.size,
+                                              size_type(end_constructed - begin_constructed)));
                 return move_elements(begin_constructed, end_constructed, smaller);
             }
